Add product screen finished

+++ app/src/main/java/com/tripiana/cabifytest/domain/model/ProductTypeModel.kt
@@ -3,5 +3,5 @@ package com.tripiana.cabifytest.domain.model
 data class ProductTypeModel(
     val code: String,
     val name: String,
-    val price: Double
+    val price: Float
 )
